auto-update: Adjust gbp conf for Debian-derived packages
Debian-derived packages are usually updated with the command 'gbp
import-dsc': it is the usual command used to import a Debian source
package in the git tree. It seems to be what's been used so far in
most/all Kali packages that are derived from Debian.

When one runs 'gbp import-dsc', gbp import the Debian sources in the
Debian branch (also called packaging branch). At the moment, this branch
is configured to be 'kali/master' by Kali's gbp.conf. And this is wrong!
What we want is for import-dsc to import the Debian sources into the
branch 'debian'. So one has to pass '--debian-branch=debian' manually.

The second thing that is wrong is that the corresponding commit will be
tagged 'kali/VERSION' (per Kali's gbp.conf), while a better name for
this tag would actually be 'debian/VERSION'. After all, it's code
imported from Debian, on the 'debian' branch, corresponding to a Debian
release.  So tagging it with 'kali/...' feels more confusing than
anything.

This commit hence attempts to fix that.

The idea is that, if ever the XSBC-Original-Maintainer field is found in
debian/control file, we assume that this package is derived from Debian,
and so we're going to use 'gbp import-dsc' to sync it with Debian. So we
configure the 'import-dsc' accordingly, that is:
1) Import the code in the 'debian' branch
2) Tag it with 'debian/%(version)s'