auto-update: Prefix scripts with a number to ensure proper ordering
The gbp.sh script actually checks for the field XSBC-Original-Maintainer
in d/control, which is itself set by the script control.sh. So we must
ensure that control.sh runs before gbp.sh. It's already the case, but
it's just luck, due to alphabetical order.

By prefixing the scripts with a number, we make it a bit more explicit
that scripts must run in a certain order.