Rework how VCS fields are handled, keep upstream VCS urls
This commit changes how VCS fields are handled.

First, the script checks if the VCS URLs seem to Kali's gitlab URLs. If
that's not the case, the fields are renamed into 'XS-<origin>-Vcs-*'.
<origin> is guessed based on the URL, it's a best effort that won't work
100% of the time. If ever we can't guess, we fallback to using
'Original' as the origin.

Then, the script adds/sets the Vcs-* fields. It enforces those fields to
be https://gitlab.com/kalilinux/packages/$SOURCE (as before, no change).
We take care of ordering (the fields should come just after the
`XS-<origin>Vcs-*` fields).

The major reason for this change is that it's useful to keep track of
the URL of the upstream package. We need this URL to maintain the
package. For everyday Kali developers it's trivial to find again this
URL, but for passer-bys, it's maybe not easy to know where the package
was forked from.

Keeping these fields should make it easier for them then.

References:
- https://www.debian.org/doc/debian-policy/ch-controlfields.html#version-control-system-vcs-fields
- https://wiki.ubuntu.com/UbuntuDevelopment/FAQ
- https://lists.ubuntu.com/archives/ubuntu-devel/2007-March/023366.html