update-mrconfig: Also checkout the debian branch (if it exists)
The command 'gbp clone' is only aware of 3 packaging branches at most:
- the packaging branch 'kali/master'
- the upstream branch 'upstream'
- the pristine-tar branch 'pristine-tar'

However, for packages that are forked from Debian, there's a 4th branch
named 'debian' with Debian's packaging. gpb is unaware of this one.

We could use 'gbp clone --all' but that seems a bit heavy-handed, and in
any case, we don't want to checkout all the potential work-in-progress
branches from various developers that exist.

So instead, with this commit we define a 'post_checkout' command that
explicitly checks for a remote branch named 'debian', and we set it up
if it exists.