update-mrconfig: Update all relevant branches
So far, 'mr update' would only run 'git pull', because we didn't
configure it to do otherwise.

Too bad, as really for packages, we don't want 'git pull' (as it updates
only the main branch), we want 'gbp pull' (as it updates the 3
branches).

Additionally, we want to update the debian branch if it exists. gbp is
not aware of this branch, so it won't do it automatically.

We could do 'gbp pull --all' but that's a bit heavy-handed. Well, it
would do the job just fine, but it will also update other tracked
branches that are unrelated, possibly work-in-progress branches that
have been force-pushed, and therefore 'gbp pull' would fail. Since we
have more than 500 repo in Kali, this unwanted side-effect would kick in
every now and then, and that would be annoying.

So instead, and following the approach from previous commit, with this
commit we selectively update the debian branch if it exits. We also make
sure to setup the local debian branch, in case it exists on the remote
but it doesn't exist locally.